Extracted aspects of the ContributorCard component into a new helper function called renderContacts. This component maps through an array of social media types that can be used as contenful content, and renders link icons from them.
